Rearrange many symbol characters.
I've been finding enter on shift a bit uncomfortable, and moving the
colon off the home row removes most of what made if feel good on the
base layer. Since its not that useful anyway, I'm shoving it back into
the symbol layer, which opens up a spot to keep the quote character.

As a consequence, we need to make room for :; keys on the symbol layer.
I haven't been happy with braces or \| keys for a while, so this is a
good chance to experiment more.