Fix Levinson Underglow
There is a new method to handle RGB underglow, so when I synced the QMK
repo my right side board was borked. For now I can enable legacy RGB
mode, with a TODO to follow the guide at
https://docs.keeb.io/docs/adding-rgb-underglow/
and rewire my keyboard to support RGB in both sides for master modes.